made modifications to the physical locations and logical rules of my smart switches.
inverter_input became AC_input and it is responsible for supplying the fridge from the grid when the grid is available.
inverter_to_fridge now turns on and off dynamically according to weather forecast and total radiance.
AC_input turns on automatically if the inverter_to_fridge is off, and wait for it to turn off before turning on if else, to protect the motor!